requirements.txt
boto==2.48.0
confusable-homoglyphs==2.0.2
decorator==4.1.2
Django==1.11.7
djangorestframework==3.7.3
django-registration==2.3
django-storages==1.6.5
easy-thumbnails==2.4.1
factory-boy==2.9.2
Faker==0.8.7
jedi==0.11.0
olefile==0.44
parso==0.1.0
pexpect==4.3.0
pickleshare==0.7.4
Pillow==4.3.0
pluggy==0.5.2
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
ptyprocess==0.5.2
py==1.5.2
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2017.3
simplegeneric==0.8.1
six==1.11.0
text-unidecode==1.1
tox==2.9.1
traitlets==4.3.2
virtualenv==15.1.0
wcwidth==0.1.7
social-auth-app-django==2.0.0
social-auth-core==1.5.0
